Feat: Add board repository

+++ src/main/java/com/example/oki/Board/repository/LikeRepository.java
@@ -9,11 +9,9 @@ public interface LikeRepository {
     // 좋아요 삭제
     void deleteLike(Long boardId, Long memberId);
 
-    // 좋아요 찾기
-
     // 게시물별 좋아요 수
-    int countByBoard(Long id);
+    Long countByBoard(Long id);
 
     // 유저별 좋아요 수
-    int countByMember(Long id);
+    Long countByMember(Long id);
 }
